readjoy/requirements.txt
cycler==0.10.0
dj-pagination==2.3.2
Django==1.11
djangorestframework==3.8.2
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.5
Pillow==5.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
reverse==0.1.0
six==1.11.0

